requirements.txt
gradio>=4.44.0
pydantic>=2.0.0
pyyaml>=6.0
httpx>=0.27.0
numpy>=1.24.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
databricks-sdk>=0.20.0
databricks-sql-connector>=3.0.0
python-multipart>=0.0.6
soundfile>=0.12.0
Pillow>=10.0.0
pytesseract>=0.3.10
pandas>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
mlflow>=2.10.0
tiktoken>=0.5.0
python-dateutil>=2.8.0
sarvamai
python-dotenv>=1.0.0
pdf2image>=1.16.0
